src/RAG-system/requirements.txt
# Core dependencies
torch>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0
FlagEmbedding>=1.2.0

# Google Gemini
google-generativeai>=0.3.0

# Utilities
requests>=2.28.0
numpy>=1.24.0
pandas>=2.0.0

# Optional: For better performance
accelerate>=0.20.0
transformers>=4.30.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0